@@ -2,126 +2,224 @@
   "name": "libp2p",
   "version": "0.36.2",
   "description": "JavaScript implementation of libp2p, a modular peer to peer network stack",
-  "leadMaintainer": "Jacob Heun <jacobheun@gmail.com>",
-  "main": "src/index.js",
-  "types": "dist/src/index.d.ts",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/libp2p/js-libp2p-interfaces/tree/master/packages/libp2p#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/libp2p/js-libp2p-interfaces.git"
+  },
+  "bugs": {
+    "url": "https://github.com/libp2p/js-libp2p-interfaces/issues"
+  },
+  "keywords": [
+    "IPFS",
+    "libp2p",
+    "network",
+    "p2p",
+    "peer",
+    "peer-to-peer"
+  ],
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=7.0.0"
+  },
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
   "typesVersions": {
     "*": {
+      "*": [
+        "*",
+        "dist/*",
+        "dist/src/*",
+        "dist/src/*/index"
+      ],
       "src/*": [
+        "*",
+        "dist/*",
         "dist/src/*",
         "dist/src/*/index"
       ]
     }
   },
   "files": [
-    "dist",
-    "src"
+    "src",
+    "dist/src",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
   ],
-  "scripts": {
-    "lint": "aegir lint",
-    "build": "aegir build",
-    "build:proto": "npm run build:proto:circuit && npm run build:proto:fetch && npm run build:proto:identify && npm run build:proto:plaintext && npm run build:proto:address-book && npm run build:proto:proto-book && npm run build:proto:peer && npm run build:proto:peer-record && npm run build:proto:envelope",
-    "build:proto:circuit": "pbjs -t static-module -w commonjs -r libp2p-circuit --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/circuit/protocol/index.js ./src/circuit/protocol/index.proto",
-    "build:proto:fetch": "pbjs -t static-module -w commonjs -r libp2p-fetch --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/fetch/proto.js ./src/fetch/proto.proto",
-    "build:proto:identify": "pbjs -t static-module -w commonjs -r libp2p-identify --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/identify/message.js ./src/identify/message.proto",
-    "build:proto:plaintext": "pbjs -t static-module -w commonjs -r libp2p-plaintext --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/insecure/proto.js ./src/insecure/proto.proto",
-    "build:proto:peer": "pbjs -t static-module -w commonjs -r libp2p-peer --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/peer-store/pb/peer.js ./src/peer-store/pb/peer.proto",
-    "build:proto:peer-record": "pbjs -t static-module -w commonjs -r libp2p-peer-record --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/record/peer-record/peer-record.js ./src/record/peer-record/peer-record.proto",
-    "build:proto:envelope": "pbjs -t static-module -w commonjs -r libp2p-envelope --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/record/envelope/envelope.js ./src/record/envelope/envelope.proto",
-    "build:proto-types": "npm run build:proto-types:circuit && npm run build:proto-types:fetch && npm run build:proto-types:identify && npm run build:proto-types:plaintext && npm run build:proto-types:address-book && npm run build:proto-types:proto-book && npm run build:proto-types:peer && npm run build:proto-types:peer-record && npm run build:proto-types:envelope",
-    "build:proto-types:circuit": "pbts -o src/circuit/protocol/index.d.ts src/circuit/protocol/index.js",
-    "build:proto-types:fetch": "pbts -o src/fetch/proto.d.ts src/fetch/proto.js",
-    "build:proto-types:identify": "pbts -o src/identify/message.d.ts src/identify/message.js",
-    "build:proto-types:plaintext": "pbts -o src/insecure/proto.d.ts src/insecure/proto.js",
-    "build:proto-types:peer": "pbts -o src/peer-store/pb/peer.d.ts src/peer-store/pb/peer.js",
-    "build:proto-types:peer-record": "pbts -o src/record/peer-record/peer-record.d.ts src/record/peer-record/peer-record.js",
-    "build:proto-types:envelope": "pbts -o src/record/envelope/envelope.d.ts src/record/envelope/envelope.js",
-    "test": "aegir test",
-    "test:ts": "aegir build --no-bundle && npm run test --prefix test/ts-use",
-    "test:node": "aegir test -t node -f \"./test/**/*.{node,spec}.js\"",
-    "test:browser": "aegir test -t browser",
-    "test:examples": "cd examples && npm run test:all",
-    "test:interop": "LIBP2P_JS=$PWD npx aegir test -t node -f ./node_modules/libp2p-interop/test/*",
-    "prepare": "npm run build",
-    "coverage": "nyc --reporter=text --reporter=lcov npm run test:node"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/libp2p/js-libp2p.git"
-  },
-  "keywords": [
-    "libp2p",
-    "network",
-    "p2p",
-    "peer",
-    "peer-to-peer",
-    "IPFS"
-  ],
-  "bugs": {
-    "url": "https://github.com/libp2p/js-libp2p/issues"
-  },
-  "homepage": "https://libp2p.io",
-  "license": "MIT",
-  "engines": {
-    "node": ">=15.0.0"
-  },
-  "browser": {
-    "nat-api": false
+  "exports": {
+    ".": {
+      "import": "./dist/src/index.js"
+    },
+    "./insecure": {
+      "import": "./dist/src/insecure/index.js"
+    }
   },
   "eslintConfig": {
     "extends": "ipfs",
+    "parserOptions": {
+      "sourceType": "module"
+    },
     "ignorePatterns": [
       "!.aegir.js",
       "test/ts-use",
       "*.d.ts"
     ]
   },
+  "release": {
+    "branches": [
+      "master"
+    ],
+    "plugins": [
+      [
+        "@semantic-release/commit-analyzer",
+        {
+          "preset": "conventionalcommits",
+          "releaseRules": [
+            {
+              "breaking": true,
+              "release": "major"
+            },
+            {
+              "revert": true,
+              "release": "patch"
+            },
+            {
+              "type": "feat",
+              "release": "minor"
+            },
+            {
+              "type": "fix",
+              "release": "patch"
+            },
+            {
+              "type": "chore",
+              "release": "patch"
+            },
+            {
+              "type": "docs",
+              "release": "patch"
+            },
+            {
+              "type": "test",
+              "release": "patch"
+            },
+            {
+              "scope": "no-release",
+              "release": false
+            }
+          ]
+        }
+      ],
+      [
+        "@semantic-release/release-notes-generator",
+        {
+          "preset": "conventionalcommits",
+          "presetConfig": {
+            "types": [
+              {
+                "type": "feat",
+                "section": "Features"
+              },
+              {
+                "type": "fix",
+                "section": "Bug Fixes"
+              },
+              {
+                "type": "chore",
+                "section": "Trivial Changes"
+              },
+              {
+                "type": "docs",
+                "section": "Trivial Changes"
+              },
+              {
+                "type": "test",
+                "section": "Tests"
+              }
+            ]
+          }
+        }
+      ],
+      "@semantic-release/changelog",
+      "@semantic-release/npm",
+      "@semantic-release/github",
+      "@semantic-release/git"
+    ]
+  },
+  "scripts": {
+    "lint": "aegir lint",
+    "build": "tsc",
+    "postbuild": "mkdir -p dist/src/circuit/pb && mkdir -p dist/src/fetch/pb && mkdir -p dist/src/identify/pb && mkdir -p dist/src/insecure/pb && cp src/circuit/pb/*.js src/circuit/pb/*.d.ts dist/src/circuit/pb && cp src/fetch/pb/*.js src/fetch/pb/*.d.ts dist/src/fetch/pb && cp src/identify/pb/*.js src/identify/pb/*.d.ts dist/src/identify/pb && cp src/insecure/pb/*.js src/insecure/pb/*.d.ts dist/src/insecure/pb",
+    "generate": "run-s generate:proto:* generate:proto-types:*",
+    "generate:proto:circuit": "pbjs -t static-module -w es6 -r libp2p-circuit --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/circuit/protocol/index.js ./src/circuit/protocol/index.proto",
+    "generate:proto:fetch": "pbjs -t static-module -w es6 -r libp2p-fetch --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/fetch/proto.js ./src/fetch/proto.proto",
+    "generate:proto:identify": "pbjs -t static-module -w es6 -r libp2p-identify --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/identify/message.js ./src/identify/message.proto",
+    "generate:proto:plaintext": "pbjs -t static-module -w es6 -r libp2p-plaintext --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/insecure/proto.js ./src/insecure/proto.proto",
+    "generate:proto-types:circuit": "pbts -o src/circuit/protocol/index.d.ts src/circuit/protocol/index.js",
+    "generate:proto-types:fetch": "pbts -o src/fetch/proto.d.ts src/fetch/proto.js",
+    "generate:proto-types:identify": "pbts -o src/identify/message.d.ts src/identify/message.js",
+    "generate:proto-types:plaintext": "pbts -o src/insecure/proto.d.ts src/insecure/proto.js",
+    "pretest": "npm run build",
+    "test": "aegir test",
+    "test:node": "npm run test -- -t node -f \"./dist/test/**/*.{node,spec}.js\" --cov",
+    "test:chrome": "npm run test -- -t browser -f \"./dist/test/**/*.spec.js\" --cov",
+    "test:chrome-webworker": "npm run test -- -t webworker -f \"./dist/test/**/*.spec.js\"",
+    "test:firefox": "npm run test -- -t browser -f \"./dist/test/**/*.spec.js\" -- --browser firefox",
+    "test:firefox-webworker": "npm run test -- -t webworker -f \"./dist/test/**/*.spec.js\" -- --browser firefox",
+    "test:electron-main": "npm run test -- -t electron-main -f \"./test/**/*.{node,spec}.js\"",
+    "test:examples": "cd examples && npm run test:all",
+    "test:interop": "LIBP2P_JS=$PWD npx aegir test -t node -f ./node_modules/libp2p-interop/test/*",
+    "release": "semantic-release"
+  },
   "dependencies": {
-    "@vascosantos/moving-average": "^1.1.0",
-    "abortable-iterator": "^3.0.0",
-    "aggregate-error": "^3.1.0",
+    "@achingbrain/nat-port-mapper": "^1.0.0",
+    "@libp2p/connection": "^1.1.4",
+    "@libp2p/crypto": "^0.22.9",
+    "@libp2p/interfaces": "^1.3.15",
+    "@libp2p/multistream-select": "^1.0.3",
+    "@libp2p/peer-id": "^1.1.8",
+    "@libp2p/peer-id-factory": "^1.0.8",
+    "@libp2p/peer-store": "^1.0.6",
+    "@libp2p/utils": "^1.0.9",
+    "@multiformats/mafmt": "^11.0.2",
+    "@multiformats/multiaddr": "^10.1.7",
+    "abortable-iterator": "^4.0.2",
+    "aggregate-error": "^4.0.0",
     "any-signal": "^3.0.0",
     "bignumber.js": "^9.0.1",
     "class-is": "^1.1.0",
     "datastore-core": "^7.0.0",
-    "debug": "^4.3.1",
-    "err-code": "^3.0.0",
-    "es6-promisify": "^7.0.0",
+    "debug": "^4.3.3",
+    "err-code": "^3.0.1",
     "events": "^3.3.0",
     "hashlru": "^2.3.0",
-    "interface-datastore": "^6.0.2",
-    "it-all": "^1.0.4",
-    "it-buffer": "^0.1.2",
-    "it-drain": "^1.0.3",
-    "it-filter": "^1.0.1",
-    "it-first": "^1.0.4",
+    "interface-datastore": "^6.1.0",
+    "it-all": "^1.0.6",
+    "it-drain": "^1.0.5",
+    "it-filter": "^1.0.3",
+    "it-first": "^1.0.6",
     "it-foreach": "^0.1.1",
-    "it-handshake": "^2.0.0",
-    "it-length-prefixed": "^5.0.2",
-    "it-map": "^1.0.4",
-    "it-merge": "^1.0.0",
-    "it-pipe": "^1.1.0",
+    "it-handshake": "^3.0.1",
+    "it-length-prefixed": "^7.0.1",
+    "it-map": "^1.0.6",
+    "it-merge": "^1.0.3",
+    "it-pipe": "^2.0.3",
     "it-sort": "^1.0.1",
-    "it-take": "^1.0.0",
-    "libp2p-crypto": "^0.21.2",
-    "libp2p-interfaces": "^4.0.0",
-    "libp2p-utils": "^0.4.0",
-    "mafmt": "^10.0.0",
+    "it-stream-types": "^1.0.4",
+    "it-take": "^1.0.2",
+    "it-to-buffer": "^2.0.2",
     "merge-options": "^3.0.4",
-    "mortice": "^2.0.1",
-    "multiaddr": "^10.0.0",
-    "multiformats": "^9.0.0",
-    "multistream-select": "^3.0.0",
+    "mortice": "^3.0.0",
+    "multiformats": "^9.6.3",
     "mutable-proxy": "^1.0.0",
-    "nat-api": "^0.3.1",
     "node-forge": "^1.2.1",
-    "p-any": "^3.0.0",
+    "p-any": "^4.0.0",
     "p-fifo": "^1.0.0",
-    "p-retry": "^4.4.0",
-    "p-settle": "^4.1.1",
-    "peer-id": "^0.16.0",
-    "private-ip": "^2.1.0",
-    "protobufjs": "^6.10.2",
+    "p-retry": "^5.0.0",
+    "p-settle": "^5.0.0",
+    "private-ip": "^2.3.3",
+    "protobufjs": "^6.11.2",
     "retimer": "^3.0.0",
     "sanitize-filename": "^1.6.3",
     "set-delayed-interval": "^1.0.0",
