@@ -1,9 +1,7 @@
-'use strict'
-
-const path = require('path')
-const execa = require('execa')
-const pDefer = require('p-defer')
-const { toString: uint8ArrayToString } = require('uint8arrays/to-string')
+import path from 'path'
+import execa from 'execa'
+import pDefer from 'p-defer'
+import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 
 function startProcess (name, args = []) {
   return execa('node', [path.join(__dirname, name), ...args], {
@@ -12,7 +10,7 @@ function startProcess (name, args = []) {
   })
 }
 
-async function test () {
+export default async function test () {
   let output1 = ''
   let output2 = ''
   let output3 = ''
@@ -90,5 +88,3 @@ async function test () {
     }
   })
 }
-
-module.exports = test
