@@ -1,34 +1,26 @@
 /* eslint-disable no-console */
-'use strict'
 
-const Libp2p = require('../../')
-const TCP = require('libp2p-tcp')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const Bootstrap = require('libp2p-bootstrap')
-
-const bootstrapers = require('./bootstrapers')
+import { createLibp2p } from '../../dist/index.js'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { Bootstrap } from '@libp2p/bootstrap'
+import bootstrapers from './bootstrappers.js'
 
 ;(async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE],
-      peerDiscovery: [Bootstrap]
-    },
-    config: {
-      peerDiscovery: {
-        bootstrap: {
-          interval: 60e3,
-          enabled: true,
-          list: bootstrapers
-        }
-      }
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncrypters: [new Noise()],
+    peerDiscoverers: [
+      new Bootstrap({
+        interval: 60e3,
+        list: bootstrapers
+      })
+    ]
   })
 
   node.connectionManager.on('peer:connect', (connection) => {
