@@ -1,30 +1,21 @@
 /* eslint-disable no-console */
-'use strict'
 
-const Libp2p = require('../../')
-const TCP = require('libp2p-tcp')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const KadDHT = require('libp2p-kad-dht')
-
-const delay = require('delay')
+import { createLibp2p } from '../../dist/src/index.js'
+import { TCP } from '@libp2p/tcp'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+import { KadDHT } from '@libp2p/kad-dht'
+import delay from 'delay'
 
 const createNode = async () => {
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     addresses: {
       listen: ['/ip4/0.0.0.0/tcp/0']
     },
-    modules: {
-      transport: [TCP],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE],
-      dht: KadDHT
-    },
-    config: {
-      dht: {
-        enabled: true
-      }
-    }
+    transports: [new TCP()],
+    streamMuxers: [new Mplex()],
+    connectionEncrypters: [new Noise()],
+    dht: KadDHT
   })
 
   await node.start()
