@@ -1,7 +1,7 @@
 'use strict'
 
-const path = require('path')
-const { waitForOutput } = require('../utils')
+import path from 'path'
+const { waitForOutput } from '../utils')
 
 async function test () {
   await waitForOutput('This message is sent on a private network', 'node', [path.join(__dirname, 'index.js')], {
