@@ -1,28 +1,22 @@
 import 'babel-polyfill'
-const Libp2p = require('libp2p')
-const WebRTCDirect = require('libp2p-webrtc-direct')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const Bootstrap = require('libp2p-bootstrap')
+import { createLibp2p } from 'libp2p'
+import WebRTCDirect from '@libp2p/webrtc-direct'
+import Mplex from '@libp2p/mplex'
+import { NOISE } from '@chainsafe/libp2p-noise'
+import { Bootstrap } from '@libp2p/bootstrap'
 
 document.addEventListener('DOMContentLoaded', async () => {
   // use the same peer id as in `listener.js` to avoid copy-pasting of listener's peer id into `peerDiscovery`
   const hardcodedPeerId = '12D3KooWCuo3MdXfMgaqpLC5Houi1TRoFqgK9aoxok4NK5udMu8m'
-  const libp2p = await Libp2p.create({
-    modules: {
-      transport: [WebRTCDirect],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE],
-      peerDiscovery: [Bootstrap]
-    },
-    config: {
-      peerDiscovery: {
-        [Bootstrap.tag]: {
-          enabled: true,
-          list: [`/ip4/127.0.0.1/tcp/9090/http/p2p-webrtc-direct/p2p/${hardcodedPeerId}`]
-        }
-      }
-    }
+  const libp2p = await createLibp2p({
+    transports: [WebRTCDirect],
+    streamMuxers: [Mplex],
+    connectionEncrypters: [NOISE],
+    peerDiscoverers: [
+      new Bootstrap({
+        list: [`/ip4/127.0.0.1/tcp/9090/http/p2p-webrtc-direct/p2p/${hardcodedPeerId}`]
+      })
+    ]
   })
 
   const status = document.getElementById('status')
