@@ -1,9 +1,8 @@
-const Libp2p = require('libp2p')
-const Bootstrap = require('libp2p-bootstrap')
-const WebRTCDirect = require('libp2p-webrtc-direct')
-const Mplex = require('libp2p-mplex')
-const { NOISE } = require('@chainsafe/libp2p-noise')
-const PeerId = require('peer-id')
+import { createLibp2p } from 'libp2p'
+import { WebRTCDirect } from '@libp2p/webrtc-direct'
+import { Mplex } from '@libp2p/mplex'
+import { Noise } from '@chainsafe/libp2p-noise'
+
 
 ;(async () => {
   // hardcoded peer id to avoid copy-pasting of listener's peer id into the dialer's bootstrap list
@@ -13,23 +12,14 @@ const PeerId = require('peer-id')
     "privKey": "CAESQAG6Ld7ev6nnD0FKPs033/j0eQpjWilhxnzJ2CCTqT0+LfcWoI2Vr+zdc1vwk7XAVdyoCa2nwUR3RJebPWsF1/I=",
     "pubKey": "CAESIC33FqCNla/s3XNb8JO1wFXcqAmtp8FEd0SXmz1rBdfy"
   })
-  const node = await Libp2p.create({
+  const node = await createLibp2p({
     peerId: hardcodedPeerId,
     addresses: {
       listen: ['/ip4/127.0.0.1/tcp/9090/http/p2p-webrtc-direct']
     },
-    modules: {
-      transport: [WebRTCDirect],
-      streamMuxer: [Mplex],
-      connEncryption: [NOISE]
-    },
-    config: {
-      peerDiscovery: {
-        [Bootstrap.tag]: {
-          enabled: false,
-        }
-      }
-    }
+    transport: [new WebRTCDirect()],
+    streamMuxers: [new Mplex()],
+    connectionEncrypters: [new Noise()]
   })
 
   node.connectionManager.on('peer:connect', (connection) => {
