@@ -1,9 +1,9 @@
 'use strict'
 
-const test1 = require('./test-1')
-const test2 = require('./test-2')
-const test3 = require('./test-3')
-const test4 = require('./test-4')
+const test1 from './test-1')
+const test2 from './test-2')
+const test3 from './test-3')
+const test4 from './test-4')
 
 async function test() {
   await test1()
